SPARKC-402: PartitionKey Indexes Pushdown Incorrect for PV3
In PV3 it is illegal to pushdown an equality predicate based on a
secondary index for a partition key while also specifying inequalities on
the token of that partition key. In PV4 it is legal. The pushdowns have
now been changed to check the PV4 and skip equality pushdowns for
partition keys if PV3 is in use.